fix: [PRL-🩹] do not use `-` at the end of the domain name and trim it to 63 chars (#171)
* fix: do not use `storybook` at the end of the domain name

It looks like we are not using `-storybook` at the end of the domain
name. Because of this it will generate wrong URL to the temploy.

* fix: replicate behavior from helm

We need to do the same thing what is `helm` doing. So we need to shrink
name of the instance to 63 characters and in case when there is a `-` at
the end we need to remove it.

https://github.com/toptal/inf-helm/blob/36a62a746fc8cd4780c35a27f7881c390ef2814f/charts/davinci-app/templates/_helpers.tpl#L16