Clean up tmp-workspaces file (#225)
The yarn install action creates a temporary tmp-workspaces.json file which stays 
in the root of the repository after the action is run and pollutes the workspace.

It breaks the actions that create PRs out of the generated changes like changesets action. 
This file is ignored in davinci, will be ignored in Talent Portal 
and all repos that implement publishing with changesets.

The file is no needed after it contributes to the cache hash.

In case of single repository (not monorepo) the file didn't exist as there was a typo 
which means the workflow works even without this file.